get typescript working and clean up a bunch of unrelated crap xD
