+++ package.json
@@ -2,6 +2,9 @@
   "name": "codegolf",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": "16.13.1"
+  },
   "dependencies": {
     "@emotion/react": "^11.7.0",
     "@emotion/styled": "^11.6.0",
@@ -47,6 +50,9 @@
     ]
   },
   "devDependencies": {
-    "prettier": "2.5.0"
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "prettier": "2.5.0",
+    "typescript": "^4.5.2"
   }
 }
