--- src/App.css
@@ -1,18 +1,0 @@
-:root {
-  --main-color: #242323;
-  --spacing-large: 15%;
-  --spacing-med: 5%;
-  --spacing-small: 3%;
-  --nav-height: 4em;
-}
-
-.App {
-  color: var(--main-color);
-}
-
-.truncate {
-  width: 100%;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
