+++ src/home/PreviewCard/PreviewCard.tsx
@@ -6,7 +6,13 @@ import CardActions from "@mui/material/CardActions";
 import Typography from "@mui/material/Typography";
 import Box from "@mui/material/Box";
 
-export const PreviewCard = ({ author, prompt, title }) => (
+export interface PreviewCardProps {
+  author: string;
+  prompt: string;
+  title: string;
+}
+
+export const PreviewCard = ({ author, prompt, title }: PreviewCardProps) => (
   <Card sx={{ width: "18rem", height: "14rem", padding: "2rem" }}>
     <Box
       sx={{
